Initial cursor left and right.
I haven't tested it shifting the screen left/right; I'll worry about that
later.

Next is cursor up and down, which is a bit more funky to get right
as I need to change the cached line and also potentially update the
whole screen setup and repaint everything.  Yes I should be scrolling
if possible instead of repainting as that'll be a LOT faster when
doing normal editing things..